fixed #5, fixd #6: create api folder and moved api related logics and types
