+++ src/App.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 import { render, screen } from '@testing-library/react'
-import { store } from './store/store'
+import { store } from './store'
 import App from './App'
 
 test('renders learn react link', () => {
